Fixed warning in CallRenderView.
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@606 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ view/CallRenderView.h
@@ -241,7 +241,7 @@ namespace view {
         inline void PropagateContext(const mmcRenderViewContext& context) {
             this->SetGpuAffinity(context.GpuAffinity);
             this->SetInstanceTime(context.InstanceTime);
-            this->SetTime(context.Time);
+            this->SetTime(static_cast<float>(context.Time));
         }
 
         /**
